ailab/requirements.txt
# Requirements for Code Quality Checker
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
httpx>=0.25.0
openai>=1.0.0
langchain-community>=0.1.0
langgraph>=0.2.0
supabase>=2.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
